Show step description on all screens
