+++ src/containers/AcceptedStepDetailScreen/index.js
@@ -21,6 +21,9 @@ class AcceptedStepDetailScreen extends Component {
 
   render() {
     const { t, step } = this.props;
+    const {
+      challenge_suggestion: { description_markdown },
+    } = step;
     const { removeStepButton, removeStepButtonText } = styles;
 
     return (
@@ -35,7 +38,7 @@ class AcceptedStepDetailScreen extends Component {
             buttonTextStyle={removeStepButtonText}
           />
         }
-        Body={null}
+        markdown={description_markdown}
         text={step.title}
         bottomButtonProps={{
           onPress: this.completeStep,
