+++ src/containers/CompletedStepDetailScreen/__tests__/CompletedStepDetailScreen.js
@@ -6,7 +6,10 @@ import { createMockNavState, renderShallow } from '../../../../testUtils';
 
 import CompletedStepDetailScreen from '..';
 
-const step = { title: 'SCOTTY' };
+const step = {
+  title: 'SCOTTY',
+  challenge_suggestion: { description_markdown: 'roge rules' },
+};
 let screen;
 
 const mockStore = configureStore([thunk]);
